Move style util functions to util.js
